pyproject.toml
[project]
name = "mcp-dev-toolkit"
version = "0.1.0"
description = "A personal MCP server with developer tools for Claude"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.2.2",
]

[project.scripts]
mcp-dev-toolkit = "server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
